Fixes #389
This makes sure dlv is used for amd64 architecture. The current release
has a few problems with darwin architecture hence the conditional build
tag.
